refactor(tangle-dapp): Move blueprint gallery to shared ui
