+++ libs/tangle-shared-ui/src/components/blueprints/BlueprintGallery/index.ts
@@ -1,0 +1,3 @@
+import BlueprintGallery from './BlueprintGallery';
+
+export default BlueprintGallery;
